LimoneIDE_Backend/requirements.txt
openai
anthropic
google-generativeai
speechrecognition
